+++ abquant/gateway/__init__.py
@@ -8,3 +8,5 @@ from .dydx.dydx_getway import DydxGateway
 from .bitmex.bitmexgateway import BitmexGateway
 from .bybit.bybit_ubc_gateway import BybitUBCGateway
 from .bybit.bybit_bbc_gateway import BybitBBCGateway
+
+from .ftx.ftx_gateway import FtxGateway