+++ abquantui/strategy_lifecycle.py
@@ -103,6 +103,12 @@ class StrategyLifecycle(ABC):
     def status(self):
         return str(self._strategy_runner.strategies.keys())
 
+    def status_live(self) -> str:
+        """
+        called by command status --live
+        """
+        return 'status_live called ' + str(time.time())
+
     @abstractmethod
     def add_init_strategy(self):
         pass